Compilation sous Linux 64 bits
Le jeu compila avec succès sur Ubuntu en 64 bits
Compilateur utilisé : g++-4.8.1 (support de -std=c++11 )
Correction nécessaire pour une compilation sur les 2 OS (windows 32 et
linux 64)
Ajouts de la bibliothèque SFML pour linux

Utilisation séparé des fichiers de projet .cbp pour les 2 OS (penser à
fusionner en utilisant différents "target" pour la compilation)

[web] corrections de fautes dans le texte à l'accueil

+++ Cpp/Scores.cpp
@@ -128,7 +128,7 @@ void Scores::loadFromFile()
         return;
     }
 
-    if (fileSize != sizeEnregistrement*nbEnregistrement+6)
+    if (fileSize != (unsigned int) (sizeEnregistrement*nbEnregistrement+6))
     {
         cout << "Erreur lors du chargement des scores" << endl;
         return;
